- Fix a lot of warnings generated by -Wshorten-64-to-32   when compiling for ia32 on Mac.
R=vegorov@google.com

Review URL: https://codereview.chromium.org//415513002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@38499 260f80e4-7a28-3924-810f-c04153c831b5

+++ runtime/vm/benchmark_test.h
@@ -71,8 +71,8 @@ class Benchmark {
   // Accessors.
   const char* name() const { return name_; }
   const char* score_kind() const { return score_kind_; }
-  void set_score(intptr_t value) { score_ = value; }
-  intptr_t score() const { return score_; }
+  void set_score(int64_t value) { score_ = value; }
+  int64_t score() const { return score_; }
   Isolate* isolate() const { return reinterpret_cast<Isolate*>(isolate_); }
 
   Dart_Isolate CreateIsolate(const uint8_t* buffer) {
@@ -98,7 +98,7 @@ class Benchmark {
   RunEntry* const run_;
   const char* name_;
   const char* score_kind_;
-  intptr_t score_;
+  int64_t score_;
   Dart_Isolate isolate_;
   Benchmark* next_;
